+++ CHANGELOG.md
@@ -2,6 +2,22 @@
 
 All notable changes to this project will be documented in this file. The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org).
 
+## [v3.0.0](https://github.com/tragiccode/tragiccode-azure_key_vault/tree/v3.0.0) (2022-08-14)
+
+[Full Changelog](https://github.com/tragiccode/tragiccode-azure_key_vault/compare/v2.1.0...v3.0.0)
+
+### Changed
+
+- \(GH-95\) Wrap hiera return in puppet sensitive data type [\#96](https://github.com/TraGicCode/tragiccode-azure_key_vault/pull/96) ([TraGicCode](https://github.com/TraGicCode))
+
+## [v2.1.0](https://github.com/tragiccode/tragiccode-azure_key_vault/tree/v2.1.0) (2022-08-08)
+
+[Full Changelog](https://github.com/tragiccode/tragiccode-azure_key_vault/compare/v2.0.2...v2.1.0)
+
+### Added
+
+- \(GH-90\) Add windows server 2022 support [\#91](https://github.com/TraGicCode/tragiccode-azure_key_vault/pull/91) ([TraGicCode](https://github.com/TraGicCode))
+
 ## [v2.0.2](https://github.com/tragiccode/tragiccode-azure_key_vault/tree/v2.0.2) (2022-08-04)
 
 [Full Changelog](https://github.com/tragiccode/tragiccode-azure_key_vault/compare/v2.0.1...v2.0.2)